+++ src/TaskFlow.Console/Models/TaskItem.cs
