+++ src/TaskFlow.Console/Services/TaskService.cs
@@ -1,9 +1,3 @@
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
-using TaskFlow.Models;
-
 namespace TaskFlow.Services
 {
     public class TaskService
