+++ src/TaskFlow.Console/TaskFlow.sln
@@ -1,0 +1,25 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio Version 17
+VisualStudioVersion = 17.14.36518.9 d17.14
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "TaskFlow.Console", "..\TaskFlow.Console\TaskFlow.Console.csproj", "{A4E7D0B7-D2E8-4921-9AD3-C95B57517E2D}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{A4E7D0B7-D2E8-4921-9AD3-C95B57517E2D}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{A4E7D0B7-D2E8-4921-9AD3-C95B57517E2D}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{A4E7D0B7-D2E8-4921-9AD3-C95B57517E2D}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{A4E7D0B7-D2E8-4921-9AD3-C95B57517E2D}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {B0C8EA25-BF76-49AB-BDAE-85AA579A941C}
+	EndGlobalSection
+EndGlobal
