+++ src/TaskFlow.Console/Utils/ConsoleHelper.cs
@@ -1,6 +1,3 @@
-using System;
-using System.Collections.Generic;
-using TaskFlow.Models;
 using TaskFlow.Services;
 
 namespace TaskFlow.Utils
